Organize App and index files
